using scons, make sucks

+++ src/networking.c
@@ -56,9 +56,9 @@ int copyfile(int fromfd, int tofd) {
   return total;
 }
 
-int sendall(int s, char *buf, int *len) {
-	int total = 0;        /* how many bytes we've sent */
-	int bytesleft = *len; /* how many we have left to send */
+int sendall(int s, char *buf, size_t *len) {
+	size_t total = 0;        /* how many bytes we've sent */
+	size_t bytesleft = *len; /* how many we have left to send */
 	int n = 0;
 
 	while(total < *len) {
